+++ v1.1.2-psi-system/app/Views/categories/index.php
@@ -41,5 +41,6 @@ $hasFilter = count(array_filter($filters, fn($v) => $v !== '')) > 0;
         </tbody>
     </table>
     </div>
+    <?php include __DIR__ . '/../partials/pagination.php'; ?>
     <?php endif; ?>
 </div>