+++ v1.1.2-psi-system/app/Views/changelogs/index.php
@@ -71,5 +71,6 @@
         </tbody>
     </table>
     </div>
+    <?php include __DIR__ . '/../partials/pagination.php'; ?>
     <?php endif; ?>
 </div>