+++ v1.1.2-psi-system/app/Views/changelogs/table.php
@@ -61,5 +61,6 @@
         </tbody>
     </table>
     </div>
+    <?php include __DIR__ . '/../partials/pagination.php'; ?>
     <?php endif; ?>
 </div>