+++ v1.1.2-psi-system/app/Views/inventory/index.php
@@ -34,5 +34,6 @@
         </tbody>
     </table>
     </div>
+    <?php include __DIR__ . '/../partials/pagination.php'; ?>
     <?php endif; ?>
 </div>